requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
celery
redis
psycopg2-binary
python-dotenv
